backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
flask==3.1.0; python_version >= '3.9'
flask-pydantic==0.12.0; python_version >= '3.7'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
pydantic==2.10.5; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.37; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
